Disable output colorization on terminals that don't support it
The current version of the colored crate ignores the capabilities of
the terminal (see https://github.com/mackwic/colored/issues/108),
which makes the outupt of probe-run unparseable in e.g. emacs
compilation mode. As a temporary workaround, this patch overrides
colorization if probe-run is called from a terminal without color
support.